requirements.txt
flask
python-dotenv
git+https://github.com/Pycord-Development/pycord